+++ src/Ma.css
@@ -2,7 +2,7 @@
 
 .back{
     /* height: 700px; */
-    background-image: linear-gradient(rgba(56, 55, 55, 0.658),rgb(56, 55, 55, 0.658)),url('./medias/minimalism_pattern_abstract_lines_geometry-163130.jpg');
+    background-image: linear-gradient(rgba(56, 55, 55, 0.658),rgb(56, 55, 55, 0.658));
     background-position: center;
     background-repeat: no-repeat;
     background-size: cover;
