+++ src/pages/index.jsx
@@ -3,4 +3,5 @@ export { default as Livraisons } from './Livraisons';
 export { default as Messagerie } from './Messagerie';
 export { default as Cours } from './Cours';
 export { default as Eleves } from './Eleves';
+export { default as Connexion } from './Connexion';
